+++ test/unit/test_A_plugin.py
@@ -5,7 +5,7 @@ Test the plugin loader mechanics
 from pii_data.types import PiiEnum
 
 from pii_extract_plg_presidio import VERSION
-from pii_extract_plg_presidio.task import PresidioTask
+from pii_extract_plg_presidio.task.task import PresidioTask
 import pii_extract_plg_presidio.plugin_loader as mod
 
 
@@ -38,8 +38,9 @@ def test30_task_descriptor():
     assert got["class"] == "PiiTask"
     assert got["task"] == PresidioTask
 
-    assert len(got["pii"]) == 5
+    assert len(got["pii"]) == 7
 
     pii_list = sorted(e["type"] for e in got["pii"])
-    assert pii_list == [PiiEnum.PERSON, PiiEnum.NORP, PiiEnum.OTHER,
-                        PiiEnum.GOV_ID, PiiEnum.GOV_ID]
+    assert pii_list == [PiiEnum.GOV_ID, PiiEnum.GOV_ID,
+                        PiiEnum.GOV_ID, PiiEnum.GOV_ID,
+                        PiiEnum.PERSON, PiiEnum.LOCATION, PiiEnum.NORP]